third_party/blink/web_tests/http/tests/security/aboutBlank/window-open-self-about-blank-expected.txt

CONSOLE MESSAGE: line 1: PASS
